This combines nickname and nicknamefix into one file
